api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.1
pydantic==2.9.2
python-dotenv==1.0.1
